% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_clusters.R
\name{my_clusters}
\alias{my_clusters}
\title{This function calculates the kmeans cluster}
\usage{
my_clusters(xy_data, cluster_value)
}
\arguments{
\item{xy_data}{list of x and y values \code{xy_data}}

\item{cluster_value}{number of clusters to calculate \code{cluster_value}}
}
\value{
cluster_list() returns a list containing calculated cluster values
}
\description{
This functions takes in the x and y variables and number of clusters
and returns a list containing the calculated cluster values by the kmeans
function.
}
\examples{
R code here showing how your function works

}
\keyword{keywords}
