% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_my_cluster.R
\name{plot_my_cluster}
\alias{plot_my_cluster}
\title{The function plots the k means cluster}
\usage{
plot_my_cluster(xy_data, cluster_data)
}
\arguments{
\item{xy_data}{list of x and y values to plot \code{xy_data}}

\item{cluster_data}{list of k-means calculated cluster data \code{cluster_data}}
}
\value{
clusterPlot returns a k-means cluster plot to render
}
\description{
This function takes in the x and y values to plot along with k-means calculated
cluster list and returns a k-means cluster plot
}
\examples{
R code here showing how your function works

}
\keyword{keywords}
